chore: RO-79 moved UpdateLineItemsService to WebHook namespace
